Removing SetTypeMappings

+++ Fluid/Filters/ColorFilters.cs
@@ -235,7 +235,7 @@ namespace Fluid.Filters
 
             var brightness = Convert.ToDouble(rgbColor.R * 299 + rgbColor.G * 587 + rgbColor.B * 114) / 1000.0;
 
-            return NumberValue.Create(Math.Round(brightness, 2));
+            return NumberValue.Create((decimal) Math.Round(brightness, 2));
         }
 
         public static ValueTask<FluidValue> ColorSaturate(FluidValue input, FilterArguments arguments, TemplateContext context)
@@ -505,7 +505,7 @@ namespace Fluid.Filters
                 var luminance2 = GetRelativeLuminance(rgbColor1);
                 var colorContrast = Math.Round((luminance1 + 0.05) / (luminance2 + 0.05), 1);
 
-                return NumberValue.Create(colorContrast);
+                return NumberValue.Create((decimal) colorContrast);
             }
         }
